Created login window.
Created file to generate login window as well as the creation of the file that holds all the logic for the login window.
The window can be fully created from the main.py file as well as testing within the login.py file by checking the file name
and seeing if its the main file being ran.