Population of base data functional.
The population of base information in the entries is functional. This can be used
to populate more than just the first base. Need to adjust the gui to display current
record and total number of records. Should probably changne the text style before comic
sans gets annoying and in the way of gui layout.