Save button functional.
Records can now be saved and will update immedately in the window. This needs to be tested for
records that aren't the first and only record. Ideally the index should remain the same unless new
records are added. New records should be added to the end of the search list and index as such and
then refresh once the window is closed.