Created VOC window.
The main menu has been changed to accomodate the addition of 'Material Recieving" button. VOC menu has been
created and linked so that progression with the Base window can continue. VOCs will need to be functional
before revisiting bases as the next step is to add VOCs to the bases and then on to creating new bases.